optimize the app structure
